# openapi.yaml
openapi: 3.0.3
info:
  title: QA Retail — API Backend
  description: >
    Contrato de la API REST de "ChileRetail", la tienda demo usada como campo de
    practica de QA. Documenta, en formato contract-first (OpenAPI), el
    comportamiento real de cada endpoint ya implementado en `src/routes/*.routes.js`:
    parametros, cuerpos de peticion, respuestas y codigos de estado (incluyendo los
    casos negativos sembrados a proposito para practicas de testing).

    Reglas transversales:
    - Todas las rutas estan bajo el prefijo `/api` (ver `servers`).
    - Las rutas marcadas con el candado requieren `Authorization: Bearer <token>`
      (JWT, expira en 2h). Sin token o token invalido/expirado -> **401**.
    - Las rutas marcadas ademas como "admin" requieren `role: admin` en el token.
      Autenticado pero sin rol admin -> **403**.
    - Toda ruta no definida en este contrato responde **404**
      `{ "error": "Ruta no encontrada" }`.
    - Cualquier excepcion no controlada responde **500**
      `{ "error": "Error interno del servidor" }`.
  version: "1.0.0"
  contact:
    name: QA Retail — Backend
servers:
  - url: http://localhost:4000/api
    description: Servidor local de desarrollo
...
security: []
paths:
...
  /products:
    get:
      tags: [products]
      summary: Lista productos, con filtros opcionales
      operationId: listProducts
      parameters:
        - name: category
          in: query
          description: Filtra por categoria exacta (case-insensitive).
          schema: { type: string }
          example: Tecnologia
        - name: search
          in: query
          description: Filtra por coincidencia parcial en el nombre (case-insensitive).
          schema: { type: string }
          example: zapatillas
      responses:
        "200":
          description: Lista de productos (vacia si no hay coincidencias).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [products]
      summary: Crea un producto (admin)
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "201":
          description: Producto creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: >
            Faltan `sku`, `name`, `price`, `stock` o `category`, o `price`/`stock`
            son negativos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [products]
      summary: Obtiene un producto por id
      operationId: getProductById
      responses:
        "200":
          description: Producto encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [products]
      summary: Actualiza un producto (admin)
      description: Todos los campos son opcionales; solo se aplican los enviados.
      operationId: updateProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
      responses:
        "200":
          description: Producto actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: "`price` o `stock` negativos."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [products]
      summary: Elimina un producto (admin)
      operationId: deleteProduct
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Producto eliminado.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /products/{id}/availability:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [products]
      summary: Disponibilidad de stock por tienda fisica
      operationId: getProductAvailability
      responses:
        "200":
          description: Stock del producto en cada tienda.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreAvailability"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ------------------------------------------------------------------ cart --
...
  /reviews/product/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema: { type: integer }
        example: 1
    get:
      tags: [reviews]
      summary: Lista las resenas de un producto, con su promedio
      operationId: getProductReviews
      responses:
        "200":
          description: Resenas del producto.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsByProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /reviews:
    post:
      tags: [reviews]
      summary: Crea o actualiza la resena del usuario sobre un producto
      description: >
        Una resena por usuario y producto: si el usuario ya habia reseniado ese
        producto, se actualiza (responde **200**); si es la primera vez, se crea
        (responde **201**).
      operationId: upsertReview
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateRequest"
      responses:
        "200":
          description: Resena existente del usuario para ese producto, actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "201":
          description: Resena creada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: >
            Faltan `productId`/`rating`, o `rating` no es un entero entre 1 y 5.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /reviews/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
        example: 1
    delete:
      tags: [reviews]
      summary: Elimina una resena propia
      operationId: deleteReview
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Resena eliminada.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La resena existe pero pertenece a otro usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # -------------------------------------------------------------- wishlist --
  /wishlist:
    get:
      tags: [wishlist]
      summary: Lista de favoritos del usuario autenticado
      operationId: getWishlist
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Productos en la wishlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /wishlist/items:
    post:
      tags: [wishlist]
      summary: Agrega un producto a la wishlist (idempotente)
      operationId: addWishlistItem
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistAddRequest"
      responses:
        "201":
          description: Wishlist actualizada (no duplica si ya existia).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "400":
          description: Falta `productId`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /wishlist/items/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema: { type: integer }
        example: 1
    delete:
      tags: [wishlist]
      summary: Quita un producto de la wishlist
      description: >
        Idempotente: responde **200** incluso si el producto no estaba en la
        wishlist (no es un error de contrato).
      operationId: removeWishlistItem
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Wishlist actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  # ------------------------------------------------------------- addresses --
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT devuelto por `POST /auth/register` o `POST /auth/login`. Payload:
        `{ id, email, role }`. Expira en 2 horas.
  responses:
    Unauthorized:
      description: Sin token (`Token no proporcionado`) o token invalido/expirado (`Token invalido o expirado`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado, pero sin rol admin (`Requiere rol admin`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Error interno no controlado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example: { error: "Mensaje descriptivo del error" }
...
    Installments:
      type: object
      properties:
        count: { type: integer, example: 12 }
        interestFree: { type: boolean, example: true }
      required: [count, interestFree]
    Product:
      type: object
      properties:
        id: { type: integer, example: 1 }
        sku: { type: string, example: SKU-001 }
        name: { type: string, example: Zapatillas Urban Runner }
        description: { type: string }
        price: { type: number, example: 45990 }
        listPrice: { type: number, example: 57490 }
        stock: { type: integer, example: 12 }
        category: { type: string, example: Calzado }
        brand: { type: string, example: UrbanStep }
        isNew: { type: boolean }
        installments:
          $ref: "#/components/schemas/Installments"
        images:
          type: array
          items: { type: string }
          description: Data URIs (SVG) de 3 variantes de color.
        videoUrl: { type: string, nullable: true }
        categoryPath:
          type: array
          items: { type: string }
          example: ["Inicio", "Calzado"]
        warrantyMonths: { type: integer, example: 12 }
        returnDays: { type: integer, example: 30 }
        imageUrl: { type: string, description: Data URI (SVG) de portada. }
      required:
        - id
        - sku
        - name
        - description
        - price
        - stock
        - category
        - imageUrl
    ProductCreateRequest:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
      required: [sku, name, price, stock, category]
    ProductUpdateRequest:
      type: object
      description: Todos los campos son opcionales; solo se aplican los enviados.
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
    StoreAvailability:
      type: object
      properties:
        storeId: { type: integer, example: 1 }
        storeName: { type: string, example: Tienda Costanera }
        qty: { type: integer, example: 2 }
      required: [storeId, storeName, qty]
...
    Review:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        userId: { type: integer }
        userName: { type: string, description: Email del autor. }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: Presente solo si la resena fue actualizada (no creada).
      required: [id, productId, userId, userName, rating, comment, createdAt]
    ReviewCreateRequest:
      type: object
      properties:
        productId: { type: integer }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
      required: [productId, rating]
    ReviewsByProductResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        average:
          type: number
          nullable: true
          description: Promedio de `rating`; `null` si no hay resenas.
        count: { type: integer }
      required: [reviews, average, count]
    WishlistResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [items]
    WishlistAddRequest:
      type: object
      properties:
        productId: { type: integer }
      required: [productId]
...